refactor(edit-profile-saga): remove user profile actions and functions and uses at update container and component level
