+++ src/store/authentication/saga.ts
@@ -15,7 +15,7 @@ import { Connectors } from '../../lib/web3';
 import { Events, getAuthChannel } from './channels';
 import { getHistory } from '../../lib/browser';
 import { completePendingUserProfile } from '../registration/saga';
-import { closeUserProfile } from '../edit-profile/saga';
+import { closeUserProfile } from '../user-profile/saga';
 
 export const currentUserSelector = () => (state) => {
   return getDeepProperty(state, 'authentication.user.data', null);
