+++ src/store/edit-profile/index.ts
@@ -3,9 +3,6 @@ import { createAction, createSlice, PayloadAction } from '@reduxjs/toolkit';
 export enum SagaActionTypes {
   EditProfile = 'profile/edit',
   FetchOwnedZIDs = 'profile/edit/fetchOwnedZIDs',
-  OpenUserProfile = 'profile/edit/openUserProfile',
-  CloseUserProfile = 'profile/edit/closeUserProfile',
-  OpenEditProfile = 'profile/edit/openEditProfile',
 }
 
 export enum State {
@@ -15,14 +12,7 @@ export enum State {
   LOADED,
 }
 
-export enum Stage {
-  None = 'none',
-  Overview = 'overview',
-  EditProfile = 'edit_profile',
-}
-
 export type EditProfileState = {
-  stage: Stage;
   errors: string[];
   state: State;
   ownedZIDs: string[];
@@ -30,7 +20,6 @@ export type EditProfileState = {
 };
 
 export const initialState: EditProfileState = {
-  stage: Stage.None,
   errors: [],
   state: State.NONE,
   ownedZIDs: [],
@@ -44,17 +33,11 @@ export const editProfile = createAction<{
 }>(SagaActionTypes.EditProfile);
 
 export const fetchOwnedZIDs = createAction(SagaActionTypes.FetchOwnedZIDs);
-export const openUserProfile = createAction(SagaActionTypes.OpenUserProfile);
-export const closeUserProfile = createAction(SagaActionTypes.CloseUserProfile);
-export const openEditProfile = createAction(SagaActionTypes.OpenEditProfile);
 
 const slice = createSlice({
   name: 'edit-profile',
   initialState,
   reducers: {
-    setStage: (state, action: PayloadAction<Stage>) => {
-      state.stage = action.payload;
-    },
     startProfileEdit: (state, _action: PayloadAction) => {
       state.errors = [];
       state.state = State.NONE;
@@ -75,5 +58,5 @@ const slice = createSlice({
   },
 });
 
-export const { setErrors, startProfileEdit, setState, setStage, setOwnedZIDs, setLoadingZIDs } = slice.actions;
+export const { setErrors, startProfileEdit, setState, setOwnedZIDs, setLoadingZIDs } = slice.actions;
 export const { reducer } = slice;
