+++ src/store/edit-profile/saga.ts
@@ -1,5 +1,5 @@
 import { call, put, select, takeLatest } from 'redux-saga/effects';
-import { SagaActionTypes, Stage, State, setErrors, setLoadingZIDs, setOwnedZIDs, setStage, setState } from '.';
+import { SagaActionTypes, State, setErrors, setLoadingZIDs, setOwnedZIDs, setState } from '.';
 import {
   editUserProfile as apiEditUserProfile,
   saveUserMatrixCredentials as apiSaveUserMatrixCredentials,
@@ -83,22 +83,7 @@ export function* fetchOwnedZIDs() {
   yield put(setLoadingZIDs(false));
 }
 
-export function* openUserProfile() {
-  yield put(setStage(Stage.Overview));
-}
-
-export function* closeUserProfile() {
-  yield put(setStage(Stage.None));
-}
-
-export function* openEditProfile() {
-  yield put(setStage(Stage.EditProfile));
-}
-
 export function* saga() {
   yield takeLatest(SagaActionTypes.EditProfile, editProfile);
   yield takeLatest(SagaActionTypes.FetchOwnedZIDs, fetchOwnedZIDs);
-  yield takeLatest(SagaActionTypes.OpenUserProfile, openUserProfile);
-  yield takeLatest(SagaActionTypes.CloseUserProfile, closeUserProfile);
-  yield takeLatest(SagaActionTypes.OpenEditProfile, openEditProfile);
 }
